requirements.txt
# Core dependencies for OMR evaluation system
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
# pillow==10.0.0
scipy==1.11.1
scikit-learn==1.3.0
matplotlib==3.7.2
imutils==0.5.4
# streamlit==1.25.0
pyyaml==6.0.1
openpyxl==3.1.2
xlsxwriter==3.1.2
plotly==5.15.0


pillow>=9.5,<11
streamlit>=1.25,<2

